Rewrite pkg structure
