+++ pkg/handler/handler.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"context"
